Reorganize API scripts with separate routes files
